Move tus related code into a separate module
