+++ cvat/apps/engine/view_utils.py
@@ -18,7 +18,7 @@ from rest_framework.response import Response
 from rest_framework.serializers import Serializer
 from rest_framework.viewsets import GenericViewSet
 
-from cvat.apps.engine.mixins import TusFile
+from cvat.apps.engine.tus import TusFile
 from cvat.apps.engine.model_utils import _ModelT
 from cvat.apps.engine.parsers import TusUploadParser
 from cvat.apps.engine.types import ExtendedRequest
